sdk/requirements.txt
# Core BNBAgent SDK dependencies
web3>=6.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# Server components  
fastapi>=0.100.0
uvicorn>=0.23.0

# Optional: IPFS support
# ipfshttpclient>=0.8.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0